Fix .u error handling in py9p.
Per the RFC, Rerror in 9p2000.u passes another field (errnum) along with
the error string ename. 9p2000.u clients (like v9fs) use this string to
recognize the error in absence of a mapping facility that relates the error
string with the error code.

This patch makes an attempt to pass error numbers along with the string
if .u is selected. For now, we don't handle all the cases and set errornum
as ERRUNDEF (~0) for the most part.

localsrv now sends a tuple of (errornum, errorstring) in cases of OSError
or IOError which turns out to be the path for most of the errors seen in
localsrv.